# Cat-Companion/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cat Care API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cats:
    get:
      operationId: listCats
      tags: [cats]
      summary: List all cats
      responses:
        "200":
          description: List of cats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cat"
    post:
      operationId: createCat
      tags: [cats]
      summary: Create a new cat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatBody"
      responses:
        "201":
          description: Created cat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
  /cats/{id}:
    get:
      operationId: getCat
      tags: [cats]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cat details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
        "404":
          description: Not found
    put:
      operationId: updateCat
      tags: [cats]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatBody"
      responses:
        "200":
          description: Updated cat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
    delete:
      operationId: deleteCat
      tags: [cats]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cats/{catId}/feedings:
    get:
      operationId: listFeedings
      tags: [feedings]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of feedings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feeding"
    post:
      operationId: logFeeding
      tags: [feedings]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogFeedingBody"
      responses:
        "201":
          description: Logged feeding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feeding"
  /feedings/{id}:
    delete:
      operationId: deleteFeeding
      tags: [feedings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /cats/{catId}/milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneBody"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /milestones/{id}:
    put:
      operationId: updateMilestone
      tags: [milestones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneBody"
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cats/{catId}/album:
    get:
      operationId: listAlbum
      tags: [photo-album]
      summary: List photo album entries for a cat
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of album entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumEntry"
    post:
      operationId: createAlbumEntry
      tags: [photo-album]
      summary: Add an entry to a cat's photo album
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumEntryBody"
      responses:
        "201":
          description: Created album entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumEntry"
  /album/{id}:
    put:
      operationId: updateAlbumEntry
      tags: [photo-album]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumEntryBody"
      responses:
        "200":
          description: Updated album entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumEntry"
    delete:
      operationId: deleteAlbumEntry
      tags: [photo-album]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cats/{catId}/medications:
    get:
      operationId: listMedications
      tags: [medications]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [medications]
      parameters:
        - name: catId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicationBody"
      responses:
        "201":
          description: Created medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /medications/{id}:
    put:
      operationId: updateMedication
      tags: [medications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicationBody"
      responses:
        "200":
          description: Updated medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
    delete:
      operationId: deleteMedication
      tags: [medications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /medications/{id}/mark-given:
    post:
      operationId: markMedicationGiven
      tags: [medications]
      summary: Mark a medication dose as given and set the next dose time
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkMedicationGivenBody"
      responses:
        "200":
          description: Updated medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
components:
  schemas:
...
    Cat:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        breed:
          type: string
          nullable: true
        birthdate:
          type: string
          nullable: true
        weightKg:
          type: number
          nullable: true
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        favouriteToy:
          type: string
          nullable: true
        favouriteFood:
          type: string
          nullable: true
        personalityTraits:
          type: string
          nullable: true
        favouriteSpot:
          type: string
          nullable: true
        favouriteSound:
          type: string
          nullable: true
        favouriteHuman:
          type: string
          nullable: true
        traitScores:
          type: string
          nullable: true
        awards:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateCatBody:
      type: object
      properties:
        name:
          type: string
        breed:
          type: string
          nullable: true
        birthdate:
          type: string
          nullable: true
        weightKg:
          type: number
          nullable: true
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        favouriteToy:
          type: string
          nullable: true
        favouriteFood:
          type: string
          nullable: true
        personalityTraits:
          type: string
          nullable: true
        favouriteSpot:
          type: string
          nullable: true
        favouriteSound:
          type: string
          nullable: true
        favouriteHuman:
          type: string
          nullable: true
        traitScores:
          type: string
          nullable: true
        awards:
          type: string
          nullable: true
      required: [name]
    UpdateCatBody:
      type: object
      properties:
        name:
          type: string
        breed:
          type: string
          nullable: true
        birthdate:
          type: string
          nullable: true
        weightKg:
          type: number
          nullable: true
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        favouriteToy:
          type: string
          nullable: true
        favouriteFood:
          type: string
          nullable: true
        personalityTraits:
          type: string
          nullable: true
        favouriteSpot:
          type: string
          nullable: true
        favouriteSound:
          type: string
          nullable: true
        favouriteHuman:
          type: string
          nullable: true
        traitScores:
          type: string
          nullable: true
        awards:
          type: string
          nullable: true
    Feeding:
      type: object
      properties:
        id:
          type: integer
        catId:
          type: integer
        fedAt:
          type: string
        foodType:
          type: string
          nullable: true
        amountGrams:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required: [id, catId, fedAt]
    LogFeedingBody:
      type: object
      properties:
        fedAt:
          type: string
        foodType:
          type: string
          nullable: true
        amountGrams:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required: [fedAt]
...
    Milestone:
      type: object
      properties:
        id:
          type: integer
        catId:
          type: integer
        title:
          type: string
        ageDescription:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        capturedAt:
          type: string
      required: [id, catId, title, capturedAt]
    CreateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        ageDescription:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        capturedAt:
          type: string
      required: [title, capturedAt]
    UpdateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        ageDescription:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        capturedAt:
          type: string
...
    AlbumEntry:
      type: object
      properties:
        id:
          type: integer
        catId:
          type: integer
        photoUrl:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        capturedAt:
          type: string
      required: [id, catId, capturedAt]
    CreateAlbumEntryBody:
      type: object
      properties:
        photoUrl:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        capturedAt:
          type: string
      required: [capturedAt]
    UpdateAlbumEntryBody:
      type: object
      properties:
        photoUrl:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        capturedAt:
          type: string
    Medication:
      type: object
      properties:
        id:
          type: integer
        catId:
          type: integer
        name:
          type: string
        dose:
          type: string
          nullable: true
        frequencyLabel:
          type: string
          nullable: true
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        isActive:
          type: boolean
        nextDoseAt:
          type: string
          nullable: true
      required: [id, catId, name, startDate, isActive]
    CreateMedicationBody:
      type: object
      properties:
        name:
          type: string
        dose:
          type: string
          nullable: true
        frequencyLabel:
          type: string
          nullable: true
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        isActive:
          type: boolean
        nextDoseAt:
          type: string
          nullable: true
      required: [name, startDate]
    UpdateMedicationBody:
      type: object
      properties:
        name:
          type: string
        dose:
          type: string
          nullable: true
        frequencyLabel:
          type: string
          nullable: true
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        isActive:
          type: boolean
        nextDoseAt:
          type: string
          nullable: true
    MarkMedicationGivenBody:
      type: object
      properties:
        nextDoseAt:
          type: string
          nullable: true
      required: []
...